Close Bronze/Silver/Gold quality-scale gaps; downgrade claim to Silver
Fixes a real bug (entity-disabled-by-default): sensor.py's add_scale_block
ignored the ENABLE_1M/1D/1MON options entirely and force-enabled every
sensor, so unchecking e.g. "Energy This Month Sensor" in config had no
effect. Now only Mains/Grid channels are force-enabled; everything else
follows the user's setting.

Structural/quality-scale changes:
- coordinator.py: real DataUpdateCoordinator subclasses (common-modules)
  instead of raw DataUpdateCoordinator(update_method=...) built inline in
  __init__.py.
- All platforms now read exclusively from ConfigEntry.runtime_data
  (VueRuntimeData) instead of hass.data[DOMAIN][entry_id] — the two were
  duplicating the same state (runtime-data rule).
- diagnostics.py added (redacts auth data, dumps device/coordinator health).
- PARALLEL_UPDATES set per platform: 0 for read-only sensor.py, 1 for
  switch.py/number.py which issue direct API writes.
- set_charger_current service is now unregistered in async_unload_entry
  (config-entry-unloading), and every user-facing exception across
  __init__.py/switch.py/number.py uses translation_domain/translation_key
  instead of raw strings or a bare re-raised requests.HTTPError
  (action-exceptions, exception-translations) — strings.json/en.json gained
  an "exceptions" section to match.
- quality_scale.yaml added, tracking every Bronze/Silver/Gold/Platinum rule
  honestly (done/todo/exempt with reasons). manifest.json's quality_scale
  claim is downgraded from the previously-unsubstantiated "platinum" to
  "silver" to match.
- README: added a Removal Instructions section, documented the config-flow
  options and the set_charger_current action, both previously undocumented.
- tests/: config flow coverage (success/invalid-auth/cannot-connect/dup
  account) and unit tests for the pure usage-sign/reset/debounce/mains-split
  math in coordinator.py, using the standard
  pytest-homeassistant-custom-component harness. NOTE: could not execute
  these in this sandbox (installing homeassistant + the test plugin via pip
  timed out) — they follow the standard patterns but need a real run in CI
  or locally to confirm they pass.

Platinum's async-dependency and inject-websession rules remain "todo" and
are not fixable in this repo: pyemvue is built on `requests`, not an async
HTTP client. Fixing that needs pyemvue itself to move to aiohttp, or this
integration to bypass pyemvue and talk to Emporia's Cognito/REST API
directly — tracked with that reasoning in quality_scale.yaml rather than
silently left off.
