+++ custom_components/emporia_vue/manifest.json
@@ -7,7 +7,7 @@
   "integration_type": "hub",
   "iot_class": "cloud_polling",
   "issue_tracker": "https://github.com/trooperthorn/ha_int_emporia/issues",
-  "quality_scale": "platinum",
+  "quality_scale": "silver",
   "requirements": ["pyemvue==0.18.9", "boto3>=1.37.1,<1.43.0"],
   "single_config_entry": true,
   "version": "2026.08.12"
