+++ custom_components/emporia_vue/number.py
@@ -20,9 +20,14 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from .charger_entity import EmporiaChargerEntity
-from .const import DOMAIN, VUE_DATA
+from .const import DOMAIN
 
 _LOGGER: logging.Logger = logging.getLogger(__name__)
+
+# This entity issues direct writes to the Emporia API (charger current).
+# Limit to one in-flight write at a time so two near-simultaneous changes
+# can't race each other against the API.
+PARALLEL_UPDATES = 1
 
 
 async def async_setup_entry(
@@ -31,13 +36,10 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up the number platform."""
-    vue: PyEmVue = hass.data[DOMAIN][config_entry.entry_id][VUE_DATA]
-    coordinator: DataUpdateCoordinator | None = hass.data[DOMAIN][config_entry.entry_id][
-        "coordinator_device_status"
-    ]
-    device_information: dict[int, VueDevice] = hass.data[DOMAIN][config_entry.entry_id][
-        "device_information"
-    ]
+    runtime = config_entry.runtime_data
+    vue: PyEmVue = runtime.vue
+    coordinator: DataUpdateCoordinator | None = runtime.coordinator_device_status
+    device_information: dict[int, VueDevice] = runtime.device_information
 
     if coordinator is None or coordinator.data is None:
         return
@@ -130,7 +132,9 @@ class EmporiaChargerCurrentNumber(EmporiaChargerEntity, NumberEntity):  # type: 
             self.async_write_ha_state()
             _LOGGER.error("Error updating charger current: %s", err)
             raise HomeAssistantError(
-                f"Failed to set Emporia charger current: {err}"
+                translation_domain=DOMAIN,
+                translation_key="charger_update_failed",
+                translation_placeholders={"error": str(err)},
             ) from err
         else:
             # Only clear the optimistic override once the coordinator's data
