+++ custom_components/emporia_vue/quality_scale.yaml
@@ -1,0 +1,90 @@
+rules:
+  # Bronze
+  action-setup: done
+  appropriate-polling: done
+  brands: todo
+  common-modules: done
+  config-flow: done
+  config-flow-test-coverage: todo
+  dependency-transparency: done
+  docs-actions: todo
+  docs-conditions:
+    status: exempt
+    comment: This integration does not provide any blueprint conditions.
+  docs-high-level-description: done
+  docs-installation-instructions: done
+  docs-removal-instructions: todo
+  docs-triggers:
+    status: exempt
+    comment: This integration does not provide any blueprint triggers.
+  entity-event-setup: done
+  entity-unique-id: done
+  has-entity-name: done
+  runtime-data: done
+  test-before-configure: done
+  test-before-setup: done
+  unique-config-entry: done
+
+  # Silver
+  action-exceptions: done
+  config-entry-unloading: done
+  docs-configuration-parameters: todo
+  docs-installation-parameters: todo
+  entity-unavailable: done
+  integration-owner: done
+  log-when-unavailable: done
+  parallel-updates: done
+  reauthentication-flow: done
+  test-coverage: todo
+
+  # Gold
+  devices: done
+  diagnostics: done
+  discovery:
+    status: exempt
+    comment: >-
+      Emporia Vue devices have no local, discoverable protocol; every device
+      lives behind a single cloud account added via config flow.
+  discovery-update-info:
+    status: exempt
+    comment: Same as discovery — no local network presence to key off of.
+  docs-data-update: todo
+  docs-examples: todo
+  docs-known-limitations: todo
+  docs-supported-devices: todo
+  docs-supported-functions: todo
+  docs-troubleshooting: todo
+  docs-use-cases: todo
+  dynamic-devices: todo
+  entity-category: todo
+  entity-device-class: done
+  entity-disabled-by-default: done
+  entity-translations: todo
+  exception-translations: done
+  icon-translations: todo
+  reconfiguration-flow: done
+  repair-issues: todo
+  stale-devices: done
+
+  # Platinum
+  async-dependency:
+    status: todo
+    comment: >-
+      Blocked upstream: pyemvue (the required client library) is built on
+      `requests`, not an async HTTP client. Every call is wrapped in
+      hass.async_add_executor_job/run_in_executor rather than awaited
+      natively. Fixing this needs pyemvue itself to move to aiohttp, or this
+      integration to bypass pyemvue and speak Emporia's Cognito/REST API
+      directly over aiohttp.
+  inject-websession:
+    status: todo
+    comment: Same blocker as async-dependency — there is no aiohttp.ClientSession in the request path to inject.
+  strict-typing:
+    status: todo
+    comment: >-
+      ConfigEntry is parameterized as ConfigEntry[VueRuntimeData] and most of
+      the codebase is fully type-hinted, but pyemvue itself ships without a
+      py.typed marker, which blocks strict mypy checking across that
+      boundary.
+
+status: silver
