+++ tests/conftest.py
@@ -1,0 +1,11 @@
+"""Shared fixtures for Emporia Vue tests."""
+
+import pytest
+
+pytest_plugins = "pytest_homeassistant_custom_component"
+
+
+@pytest.fixture(autouse=True)
+def auto_enable_custom_integrations(enable_custom_integrations):
+    """Enable custom integrations for every test in this suite."""
+    yield
